Version 2.93+ working on additional bacthProcess features

+++ src/it/unitn/ing/rista/diffr/Sample.java
@@ -1791,7 +1791,7 @@ public class Sample extends Maincat {
     }
     double toLambda = 0.0f;
     if (isTOF) {
-      toLambda = (double) (2.0 * MoreMath.sind(Math.abs(tilting_angles[0])));
+      toLambda = (2.0 * MoreMath.sind(Math.abs(tilting_angles[0])));
     }
 
     double x, y, r;
